requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
python-socketio>=5.10.0
eventlet>=0.33.0
ultralytics>=8.0.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
werkzeug>=3.0.0
